backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.3
geopandas==0.14.1
Pillow==10.1.0
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1